deal with no agent-install.crt, exit is docker is not installed

+++ agent-install/agent-install.sh
@@ -453,7 +453,7 @@ function create_config() {
 				-e "s~^HZN_DEVICE_ID=[^ ]*~HZN_DEVICE_ID=${DEVICE_ID}~g" \
 				-e "s~^HZN_FSS_CSSURL=[^ ]*~HZN_FSS_CSSURL=${HZN_FSS_CSSURL}~g" /etc/default/horizon
 		else
-			if [[ ${CERTIFICATE:0:1} != "/" ]]; then
+			if [[ ${CERTIFICATE:0:1} != "/" ]] && [[ -f "$CERTIFICATE" ]]; then
 				sudo mv $CERTIFICATE /etc/horizon/agent-install.crt
 				ABS_CERTIFICATE=/etc/horizon/agent-install.crt
 			else
@@ -485,7 +485,7 @@ function install_macos() {
 
     sudo security add-trusted-cert -d -r trustRoot -k /Library/Keychains/System.keychain ${PACKAGES}/${MAC_PACKAGE_CERT}
     { set +x; } 2>/dev/null
-	if [[ "$CERTIFICATE" != "" ]]; then
+	if [[ "$CERTIFICATE" != "" ]] && [[ -f "$CERTIFICATE" ]]; then
 		log_info "Configuring an edge node to trust the ICP certificate ..."
 		set -x
 		sudo mv $CERTIFICATE /private/etc/horizon/agent-install.crt
@@ -584,7 +584,7 @@ function install_macos() {
 		printf "HZN_EXCHANGE_URL=${HZN_EXCHANGE_URL} \nHZN_FSS_CSSURL=${HZN_FSS_CSSURL} \
 			\nHZN_DEVICE_ID=${DEVICE_ID}"  | sudo tee "$HZN_CONFIG"
 	else
-		if [[ ${CERTIFICATE:0:1} != "/" ]]; then
+		if [[ ${CERTIFICATE:0:1} != "/" ]] && [[ -f "$CERTIFICATE" ]]; then
 			#ABS_CERTIFICATE=$(pwd)/${CERTIFICATE}
 			sudo mv $CERTIFICATE /private/etc/horizon/agent-install.crt
 			ABS_CERTIFICATE=/private/etc/horizon/agent-install.crt
@@ -670,11 +670,7 @@ function install_linux(){
     if command -v docker >/dev/null 2>&1; then
         log_info "docker found"
     else
-        log_info "docker not found, installing it..."
-        set -x
-        apt install -y docker-ce docker-ce-cli containerd.io
-        { set +x; } 2>/dev/null
-        log_info "docker installed"
+        log_info "docker not found, install before continuing..."
     fi
 
     if [[ -n "$PKG_APT_REPO" ]]; then
